[Devops] fix backend flow (#166)
Fix backend flow; in the end, the problem was new adjust copy&paste from the CLI flow. Also, update CLI flow to match backend.

I commented out most of devops test flow, to avoid spending on agents. But keep the file, so it's easy to use it in the future.